requirements.txt
aniso8601==8.0.0
attrs==20.2.0
click==8.1.3
Flask==1.1.2
Flask-Cors==3.0.10
flask-restx==0.5.1
Flask-Script==2.0.6
importlib-metadata==5.0.0
itsdangerous==1.1.0
Jinja2==3.0.3
jsonschema==4.16.0
MarkupSafe==2.1.1
pyrsistent==0.19.1
python-dotenv==0.21.0
pytz==2022.5
six==1.16.0
Werkzeug==0.16.0
zipp==3.10.0
